Refactor the Twisted consumer
Move the consuming code from the protocol to the consumer, and adjust
tests.
