+++ fedora_messaging/tests/unit/twisted/utils.py
@@ -21,6 +21,8 @@ from unittest import mock
 import pika
 from twisted.internet import defer
 
+from fedora_messaging.twisted.protocol import FedoraMessagingProtocolV2
+
 
 class MockChannel(mock.Mock):
     """A mock object with Channel-specific methods that return Deferreds."""
@@ -54,3 +56,21 @@ class MockChannel(mock.Mock):
         self.basic_consume = mock.Mock(
             side_effect=lambda **kw: defer.succeed((self.queue_object, "consumer-tag"))
         )
+
+
+class MockProtocol(FedoraMessagingProtocolV2):
+    """A Protocol object that mocks the underlying channel and impl."""
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._impl = mock.Mock(name="_impl")
+        self._impl.is_closed = True
+        self._channel = MockChannel(name="_channel")
+        self.channel = mock.Mock(
+            name="channel", side_effect=lambda: defer.succeed(self._channel)
+        )
+
+    def _register_consumer(self, consumer):
+        consumer._protocol = self
+        consumer._channel = self._channel
+        self._consumers[consumer.queue] = consumer
