+++ news/PR249.api
@@ -1,1 +1,3 @@
-Drop the Twisted classes that had been flagged as deprecated. Drop the deprecated ``Message._body`` property.
+Drop the Twisted classes that had been flagged as deprecated.
+Drop the deprecated ``Message._body`` property.
+Refactor the consuming code into the ``Consumer`` class.
